+++ src/Nakama/Ninja.WebSockets/Internal/WebSocketFrame.cs
@@ -1,4 +1,5 @@
-﻿using System.Net.WebSockets;
+﻿using System;
+using System.Net.WebSockets;
 
 namespace Nakama.Ninja.WebSockets.Internal
 {
@@ -14,18 +15,20 @@ namespace Nakama.Ninja.WebSockets.Internal
 
         public string CloseStatusDescription { get; private set; }
 
-        public WebSocketFrame(bool isFinBitSet, WebSocketOpCode webSocketOpCode, int count)
+        public ArraySegment<byte> MaskKey { get; private set; }
+
+        public WebSocketFrame(bool isFinBitSet, WebSocketOpCode webSocketOpCode, int count, ArraySegment<byte> maskKey)
         {
             IsFinBitSet = isFinBitSet;
             OpCode = webSocketOpCode;
             Count = count;
+            MaskKey = maskKey;
         }
 
-        public WebSocketFrame(bool isFinBitSet, WebSocketOpCode webSocketOpCode, int count, WebSocketCloseStatus closeStatus, string closeStatusDescription) : this(isFinBitSet, webSocketOpCode, count)
+        public WebSocketFrame(bool isFinBitSet, WebSocketOpCode webSocketOpCode, int count, WebSocketCloseStatus closeStatus, string closeStatusDescription, ArraySegment<byte> maskKey) : this(isFinBitSet, webSocketOpCode, count, maskKey)
         {
             CloseStatus = closeStatus;
             CloseStatusDescription = closeStatusDescription;
         }
-
     }
 }