+++ src/Nakama/Ninja.WebSockets/PingPongManager.cs
@@ -59,15 +59,12 @@ namespace Nakama.Ninja.WebSockets
         /// </param>
         /// <param name="cancellationToken">The token used to cancel a pending ping send AND the automatic sending of ping messages
         /// if keepAliveInterval is positive</param>
-        public PingPongManager(Guid guid, WebSocket webSocket, TimeSpan keepAliveInterval, CancellationToken cancellationToken)
+        public PingPongManager(Guid guid, WebSocket webSocket, TimeSpan keepAliveInterval,
+            CancellationToken cancellationToken)
         {
             var webSocketImpl = webSocket as WebSocketImplementation;
-            _webSocket = webSocketImpl;
-            if (_webSocket == null)
-            {
-                throw new InvalidCastException(
-                    "Cannot cast WebSocket to an instance of WebSocketImplementation. Please use the web socket factories to create a web socket");
-            }
+            _webSocket = webSocketImpl ?? throw new InvalidCastException(
+                "Cannot cast WebSocket to an instance of WebSocketImplementation. Please use the web socket factories to create a web socket");
             _guid = guid;
             _keepAliveInterval = keepAliveInterval;
             _cancellationToken = cancellationToken;
@@ -114,7 +111,9 @@ namespace Nakama.Ninja.WebSockets
 
                     if (_pingSentTicks != 0)
                     {
-                        await _webSocket.CloseAsync(WebSocketCloseStatus.NormalClosure, $"No Pong message received in response to a Ping after KeepAliveInterval {_keepAliveInterval}", _cancellationToken);
+                        await _webSocket.CloseAsync(WebSocketCloseStatus.NormalClosure,
+                            $"No Pong message received in response to a Ping after KeepAliveInterval {_keepAliveInterval}",
+                            _cancellationToken);
                         break;
                     }